MSI-B150m-gaming-pro
Clover is fully adapted for this motherboard, the system works just fine, this is a very good motherboard for Mac OS (albeit a great rarity for ISI), in general, only the latest BIOS version is needed for this config. On the official website there is the necessary BIOS. I assembled the SSDT for it, made the kekt legacy for USB (NEVER UPDATE IT) this KEXT for all life, in general, that's how it is.